+++ eox_core/__init__.py
@@ -1,4 +1,4 @@
 """
 Init for main eox-core app
 """
-__version__ = '3.3.0'
+__version__ = '3.4.0'